merge person-contract-beans-logic into main branch

See merge request jeewa-2026-bravo/jeewa-2026-bravo!7

+++ BRAVO-TSS-ejb/src/java/tss/logic/impl/PersonLogicImp.java
@@ -2,7 +2,11 @@ package tss.logic.impl;
 
 import jakarta.ejb.EJB;
 import jakarta.ejb.Stateless;
+import java.util.ArrayList;
+import java.util.List;
 import tss.dao.PersonDao;
+import tss.dto.ContractDTO;
+import tss.dto.PersonDTO;
 import tss.entity.Person;
 import tss.entity.Role;
 import tss.logic.PersonLogic;
@@ -15,7 +19,7 @@ public class PersonLogicImp implements PersonLogic {
     private PersonDao personDao;
 
     @Override
-    public Person createPerson(String firstName, String lastName, String emailAddress, boolean consent, String password, Role role) {
+    public PersonDTO createPerson(String firstName, String lastName, String emailAddress, boolean consent, String password, Role role) {
         Person person = new Person();
         person.setFirstName(firstName);
         person.setLastName(lastName);
@@ -26,11 +30,11 @@ public class PersonLogicImp implements PersonLogic {
 
         personDao.createPerson(person);
 
-        return person;
+        return personDto(person);
     }
 
     @Override
-    public Person findPerson(Long id) {
+    public PersonDTO findPerson(Long id) {
 
         Person person = personDao.findPersonById(id);
 
@@ -40,16 +44,60 @@ public class PersonLogicImp implements PersonLogic {
             );
         }
 
-        return person;
+        return personDto(person);
     }
 
     @Override
-    public Person updatePerson(Person person) {
-        return personDao.updatePerson(person);
+    public PersonDTO updatePerson(PersonDTO dto) {
+    Person person = personDao.findPersonById(dto.getId());  
+    
+    
+    if (person == null) {
+        throw new IllegalArgumentException("No Person found with id: " + dto.getId());
+    }
+    person.setFirstName(dto.getFirstName());
+    person.setLastName(dto.getLastName());
+    person.setEmailAddress(dto.getEmailAddress());
+    person.setConsent(dto.isConsent());
+    person.setRole(dto.getRole());
+    return personDto(personDao.updatePerson(person));
     }
 
     @Override
-    public void deletePerson(Person person) {
-        personDao.deletePerson(person);
+    public void deletePerson(PersonDTO dto) {
+    Person person = personDao.findPersonById(dto.getId());
+    if (person == null) {
+        throw new IllegalArgumentException("No Person found with id: " + dto.getId());
     }
+    personDao.deletePerson(person);
+}
+
+    @Override
+    public List<PersonDTO> findAllPersons() {
+          List<Person> persons = personDao.findAllPersons();
+    return toDtoList(persons);
+    }
+    
+    private List<PersonDTO> toDtoList(List<Person> persons) {
+    List<PersonDTO> dtoList = new ArrayList<>();
+
+    for (Person p : persons) {
+        dtoList.add(personDto(p));
+    }
+
+    return dtoList;
+}
+    
+  private PersonDTO personDto(Person p) {
+    PersonDTO dto = new PersonDTO();
+    dto.setId(p.getId());
+    dto.setUuid(p.getUuid());
+    dto.setFirstName(p.getFirstName());
+    dto.setLastName(p.getLastName());
+    dto.setEmailAddress(p.getEmailAddress());
+    dto.setConsent(p.isConsent());
+    dto.setRole(p.getRole());
+    return dto;
+}
+
 }
